2021.06.25
1. 대시보드 상세 페이지 작업

+++ delivery-dispatch-core/src/main/java/kr/co/cntt/core/model/dashboard/RankInfo.java
@@ -14,6 +14,6 @@ import lombok.ToString;
 @ToString
 public class RankInfo extends Common implements Dto {
     private String storeName;               // 매장 이름
-    private float value;                     // 정상 주문 건수
-    private int achievementRate;            // 달성률
+    private Float value;                     // 정상 주문 건수
+    private Float achievementRate;            // 달성률
 }
